fixed an issue with the route
